pmproxy: fix redisfmt() reply len sizeof mismatch on 32bit platforms
Redis sdscatfmt '%U' expects an unsigned long long argument,
but in redisfmt(), reply->len is sizeof_t, which is 32bits
on 32bit platforms. Cast the argument to uint64_t, which for
all practical purposes in an unsigned long long on all platforms.

Reference https://github.com/redis/hiredis/blob/master/sds.c line 640

qa/1602 and 1214 on 32bit platforms are now passing. 64bit platforms
are unaffected.

+++ src/pmproxy/src/redis.c
@@ -50,7 +50,7 @@ redisfmt(redisReply *reply)
 
     switch (reply->type) {
     case REDIS_REPLY_STRING:
-	return sdscatfmt(command, "$%U\r\n%s\r\n", reply->len, reply->str);
+	return sdscatfmt(command, "$%U\r\n%s\r\n", (uint64_t)reply->len, reply->str);
     case REDIS_REPLY_ARRAY:
     case REDIS_REPLY_MAP:
     case REDIS_REPLY_SET:
